chore: migrate database from mysql to sqlite (#15)

+++ server/src/snapshot/entity/snapshot.entity.ts
@@ -5,16 +5,16 @@ export class Snapshot {
   @PrimaryGeneratedColumn()
   id!: number;
 
-  @Column({ type: 'double', nullable: true })
+  @Column({ type: 'float', nullable: true })
   download: number | null = null;
 
-  @Column({ type: 'double', nullable: true })
+  @Column({ type: 'float', nullable: true })
   upload: number | null = null;
 
-  @Column({ type: 'double', nullable: true })
+  @Column({ type: 'float', nullable: true })
   ping: number | null = null;
 
-  @Column({ type: 'double', nullable: true })
+  @Column({ type: 'float', nullable: true })
   loss: number | null = null;
 
   @Column({ type: 'varchar', length: 255, nullable: true })
